+++ src/screens/UsageInstructions/index.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { View, Text } from 'react-native';
-import UserInfo from '_components/user-settings/';
+import UserInfo from '_components/user-info/';
 import { faTools, faAngleLeft } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { StyledView } from '../../styles/shared/StyledView';
