+++ src/screens/UserSettings/index.js
@@ -2,14 +2,14 @@ import React, { useContext } from 'react';
 import { View, Text, TouchableOpacity } from 'react-native';
 import { Context as AuthContext } from 'state/AuthContext';
 import { StyledView } from 'styles/shared/StyledView';
-import UserInfo from '_components/user-settings/';
+import UserInfo from '_components/user-info';
 import { useTranslation } from 'react-i18next';
 import { styles } from './UserSettings.styles';
 
 export default function UserSettings({ navigation }) {
   const { t } = useTranslation();
   const { state, signout } = useContext(AuthContext);
-  const { username } = state;
+  const { username, token } = state;
   const handleSignOut = () => {
     signout();
   };
@@ -22,7 +22,7 @@ export default function UserSettings({ navigation }) {
             {t('common.how-to-use')}
           </Text>
         </TouchableOpacity>
-        {state.token ? (
+        {token ? (
           <>
             <TouchableOpacity
               onPress={() => navigation.navigate('Change Password')}>
@@ -45,7 +45,7 @@ export default function UserSettings({ navigation }) {
           </>
         ) : null}
       </View>
-      {!state.token ? (
+      {!token ? (
         <TouchableOpacity onPress={() => navigation.navigate('Login')}>
           <Text style={styles.button}>{t('common.login')}</Text>
         </TouchableOpacity>
